# APIMercadoSubastas/Docs/swagger_final.yaml
openapi: 3.0.0
info:
  title: API de Mercado Subastas
  description: API para la aplicacion de subastas de la materia Desarrollo de Aplicaciones 1
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
...
paths:
...
  /auth/login:
    post:
      tags: [Auth y Registro]
      summary: Iniciar sesión
      description: Autentica al usuario con mail y contraseña.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mail, contraseña]
              properties:
                mail:
                  type: string
                  format: email
                  example: "juan.perez@ejemplo.com"
                contraseña:
                  type: string
                  format: password
                  example: "Xk9#mP2q"
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Usuario no autenticado
        '403':
          description: Usuario autenticado pero sin permisos
        '405':
          description: Fallo interno del servidor
...
  /auth/cambiar-clave:
    put:
      tags: [Auth y Registro]
      summary: Cambiar contraseña
      description: Cambio de clave luego de ser aceptado en la plataforma
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [claveActual, nuevaClave, confirmarNuevaClave]
              properties:
                claveActual:
                  type: string
                  format: password
                  description: Contraseña actual (puede ser la temporal recibida por mail)
                  example: "Xk9#mP2q"
                nuevaClave:
                  type: string
                  format: password
                  minLength: 8
                  description: Mínimo 8 caracteres, al menos un número y un símbolo
                  example: "MiClave456@"
                confirmarNuevaClave:
                  type: string
                  format: password
                  example: "MiClave456@"
      responses:
        '200':
          description: Contraseña actualizada
          content:
            application/json:
              schema:
                type: string
                example: Contraseña actualizada con exito
        '400':
          description: La clave actual es incorrecta o las nuevas no coinciden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Usuario no autenticado
        '403':
          description: Usuario autenticado pero sin permisos
        '405':
          description: Fallo interno del servidor

  # ── REGISTRO ──────────────────────────────────────────────────────────────
  /auth/registro/iniciar:
    post:
      tags: [Auth y Registro]
      summary: Paso 1 — Enviar datos personales
      description: |
        El usuario carga sus datos. Se crea el registro en `personas` y
        `persona_detalle`. El estado queda pendiente de verificación por
        un empleado de la empresa.

        La app debe pasar a la pantalla de espera (paso 2).

        Tablas que se escriben: `personas`, `persona_detalle`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, apellido, documento, mail, direccion, pais]
              properties:
                nombre:
                  type: string
                  example: "Juan"
                apellido:
                  type: string
                  example: "Pérez"
                documento:
                  type: string
                  description: Número de documento (DNI, pasaporte, etc.)
                  example: "35123456"
                mail:
                  type: string
                  format: email
                  example: "juan.perez@ejemplo.com"
                direccion:
                  type: string
                  description: Domicilio legal
                  example: "Av. Corrientes 1234, CABA"
                pais:
                  type: integer
                  description: Número del país (FK tabla `paises`)
                  example: 1
      responses:
        '201':
          description: Registro inicial creado. La app muestra la pantalla de espera.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: "Datos recibidos. Recibirás un mail con tu contraseña cuando seas verificado."
                  personaId:
                    type: integer
                    example: 42
        '401':
          description: Usuario no autenticado
        '403':
          description: Usuario autenticado pero sin permisos
        '405':
          description: Fallo interno del servidor
        '409':
          description: Ya existe un usuario con ese mail o número de documento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                codigo: 409
                mensaje: "Ya existe un usuario registrado con ese correo electrónico"
...
  /mediosPago:
    get:
      tags: [Medios de Pago]
      summary: Obtener todos los medios de pago del cliente autenticado
      description: "Tabla consultada: `mediosDePago` (tabla nueva)"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de medios de pago
          content:
            application/json:
              schema:
                type: object
                properties:
                  tieneMedioPagoVerificado:
                    type: boolean
                    description: Si al menos uno está verificado, el usuario puede pujar
                    example: false
                  medios:
                    type: array
                    items:
                      $ref: '#/components/schemas/MedioPago'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /mediosPago/cuenta-bancaria:
    post:
      tags: [Medios de Pago]
      summary: Agregar cuenta bancaria
      description: |
        Nacional o del exterior. Queda en estado `pendiente` hasta verificación.
        Tabla que se escribe: `mediosDePago`
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaBancariaInput'
      responses:
        '201':
          description: Cuenta registrada (pendiente de verificación)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedioPago'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /mediosPago/tarjeta:
    post:
      tags: [Medios de Pago]
      summary: Agregar tarjeta de crédito o débito
      description: |
        Acepta Visa, Mastercard y AMEX, nacionales o internacionales.
        Para subastas en dólares solo se aceptan tarjetas internacionales.
        Tabla que se escribe: `mediosDePago`
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TarjetaInput'
      responses:
        '201':
          description: Tarjeta registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedioPago'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /mediosPago/cheque:
    post:
      tags: [Medios de Pago]
      summary: Registrar cheque certificado
      description: |
        Debe ser entregado y verificado presencialmente ANTES del inicio de la subasta.
        Las compras acumuladas no pueden superar el monto del cheque.
        Tabla que se escribe: `mediosDePago`
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChequeCertificadoInput'
      responses:
        '201':
          description: Cheque registrado (pendiente de verificación presencial)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedioPago'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /mediosPago/{id}:
    get:
      tags: [Medios de Pago]
      summary: Obtener detalle de un medio de pago
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Detalle del medio de pago
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedioPago'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El medio de pago no pertenece al usuario autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Medios de Pago]
      summary: Actualizar descripción de un medio de pago
      description: Solo se permite modificar la descripción. Para cambiar datos bancarios, eliminar y crear uno nuevo.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                descripcion:
                  type: string
                  example: "Mi tarjeta Visa principal"
      responses:
        '200':
          description: Actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedioPago'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El medio de pago no pertenece al usuario autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Medios de Pago]
      summary: Eliminar un medio de pago
      description: |
        No se puede eliminar si:
        - Es el único medio de pago registrado
        - Está asociado a una subasta activa o compra pendiente de pago
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Eliminado correctamente
        '400':
          description: No se puede eliminar por restricciones de negocio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                codigo: 400
                mensaje: "No podés eliminar tu único medio de pago. Agregá otro primero."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El medio de pago no pertenece al usuario autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Home y Catalogo ────────────────────────────────────────────────────────────────
...
  /subastas/{subastaId}/catalogo:
    get:
      summary: Obtener catálogo de una subasta
      description: Devuelve el catálogo de productos asociados a una subasta específica.
      tags: [Home y Catalogo]
      parameters:
        - name: subastaId
          in: path
          required: true
          description: Identificador de la subasta
          schema:
            type: integer
            example: 12
      responses:
        '200':
          description: Catálogo obtenido correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Producto'
        '404':
          description: Subasta no encontrada
        '500':
          description: Error interno del servidor
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtenido en POST /auth/login
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
      example: 1
  responses:
    Unauthorized:
      description: No autenticado o token inválido/expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            codigo: 401
            mensaje: "Token inválido o sesión expirada"
...
  schemas:
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT para incluir en Authorization header
          example: "eyJhbGciOiJIUzI1NiJ9..."
        usuario:
          $ref: '#/components/schemas/UsuarioResumen'
...
    ValidationErrorResponse:
      type: object
      properties:
        codigo:
          type: integer
          example: 422
        errores:
          type: array
          items:
            type: object
            properties:
              campo:
                type: string
                example: "mail"
              mensaje:
                type: string
                example: "El formato del correo no es válido"
    Error:
      type: object
      properties:
        codigo:
          type: integer
          example: 404
        mensaje:
          type: string
          example: "Recurso no encontrado"
...